restructure files in client folder Relates #52
