+++ client/components/Categories.tsx
