+++ client/components/Navbar.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { Dialog, DialogPanel } from "@headlessui/react";
 import { Bars3Icon, XMarkIcon } from "@heroicons/react/24/outline";
 import Image from "next/image";
-import logo from "../../assets/logo.png"
+import logo from "../assets/logo.png"
 
 const navigation = [
   { name: "Product", href: "#" },
