+++ client/components/Quickview.tsx
